Refactor task management and remove 'task_manager'
The 'task_manager' has been removed and its functionalities have been integrated into 'task'. This simplifies the architecture, making task management more straightforward and having less dependency. The movement operation has been replaced with memcpy for better memory handling and execution control flow.
