+++ kernel/task/CMakeLists.txt
@@ -1,7 +1,6 @@
 set(TASK_SOURCE_FILES
-    context_switch.asm
-    task_manager.cpp
-    task.cpp
+        context_switch.asm
+        task.cpp
 )
 
 add_library(UchosTask ${TASK_SOURCE_FILES})
