+++ kernel/timers/timer.cpp
@@ -3,7 +3,7 @@
 #include "../memory/slab.hpp"
 #include "../system_event.hpp"
 #include "../system_event_queue.hpp"
-#include "../task/task_manager.hpp"
+#include "../task/task.hpp"
 #include "../types.hpp"
 #include <algorithm>
